refactor: mui theme 선언 모아보기
